dashboard/requirements.txt
Flask
Flask-Admin
Flask-Security
Flask-SQLAlchemy
wtforms
psycopg2
SQLAlchemy-serializer
jsonschema
email_validator
bcrypt
